UAT VPN

+++ src/main/java/com/mfs/merchantQR/repo/TblMenuRepo.java
@@ -3,5 +3,5 @@ package com.mfs.merchantQR.repo;
 import com.mfs.merchantQR.model.TblMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TblMenuRepo extends JpaRepository<TblMenu,Integer> {
+public interface TblMenuRepo extends JpaRepository<TblMenu,Long> {
 }
