+++ src/main/java/com/mfs/merchantQR/repo/TblRoleRightRepo.java
@@ -3,5 +3,5 @@ package com.mfs.merchantQR.repo;
 import com.mfs.merchantQR.model.TblRoleRight;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TblRoleRightRepo extends JpaRepository<TblRoleRight,Integer> {
+public interface TblRoleRightRepo extends JpaRepository<TblRoleRight,Long> {
 }
