+++ src/main/java/com/mfs/merchantQR/repo/TblUserRoleRepo.java
@@ -3,5 +3,5 @@ package com.mfs.merchantQR.repo;
 import com.mfs.merchantQR.model.TblUserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TblUserRoleRepo extends JpaRepository<TblUserRole,Integer> {
+public interface TblUserRoleRepo extends JpaRepository<TblUserRole,Long> {
 }
